% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova3_ss.R
\name{anova3_ss}
\alias{anova3_ss}
\title{One way ANOVA sample size calculation for omnibus F test with 3 group levels}
\usage{
anova3_ss(
  means,
  sds,
  alphalevel,
  powerlevel,
  groupratios = NULL,
  Nmin = NULL,
  Nmax = NULL,
  upperbound = 500
)
}
\arguments{
\item{means}{A vector of group means.}

\item{sds}{A vector of group standard deviations. This function allows unequal standard deviations between groups.}

\item{alphalevel}{The significance level.}

\item{powerlevel}{The level of power to achieve as a decimal.}

\item{groupratios}{A 3 element vector of allocation ratios to each group. Default is a 1:1:1 ratio, supplied as c(1,1,1).}

\item{Nmin}{The minimum total sample size N to search through. Default start is 30.}

\item{Nmax}{The maximum total sample size N to search through. Default ending is 300.}

\item{upperbound}{The largest integer to multiply the allocation ratios by. Default is 500 meaning if there is even allocation, this will allow a search of up to 500 people per group.}
}
\value{
Returns a data frame with sample sizes for each group indicated in n1, n2, and n3, actual power, and any notes from the computation.
}
\description{
This function searches through different group sample sizes until the specified power level is achieved.
It allows for unequal standard deviations between groups and unequal group allocation.
This function is a wrapper for the 'anova1f_3' function in the 'pwr2ppl' package.
}
\examples{
# Sample Size for 3 groups with unequal allocation at 0.05 significance and 80\% power
# explores range itself from N = 30 to 300
anova3_ss(means = c(5, 10, 12), sds = c(10, 8, 10), alphalevel = 0.05,
powerlevel = 0.8, groupratios = c(1, 1, 3),
Nmin = NULL, Nmax = NULL)
}
