% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova4_ss_wrapper.R
\name{anova4_ss_wrapper}
\alias{anova4_ss_wrapper}
\title{One way ANOVA sample size calculation for omnibus F test with 4 group levels}
\usage{
anova4_ss_wrapper(
  means,
  sds,
  alphalevel,
  powerlevel,
  groupweights,
  Nmin = NULL,
  Nmax = NULL,
  Nchange = NULL
)
}
\arguments{
\item{means}{A vector of group means.}

\item{sds}{A vector of group standard deviations. This function allows unequal standard deviations between groups.}

\item{alphalevel}{The significance level.}

\item{powerlevel}{The level of power to achieve.}

\item{groupweights}{A vector of allocation to each group in terms of decimals. This needs to sum to 1.}

\item{Nmin}{The minimum sample size N to search through. If not supplied the function starts at 30.}

\item{Nmax}{The maximum sample size N to search through. If not supplied the function ends at 150.}

\item{Nchange}{The initial sample size step to use in the rough search. If not supplied the function searches by increments of 30 people.}
}
\value{
Returns a data frame with sample sizes for each group indicated in n1, n2, n3, and n4, actual power, and any notes from the computation.
}
\description{
This function searches through different group sample sizes until the specified power level is achieved.
It allows for unequal standard deviations between groups and unequal group allocation.
}
\examples{
# explores range
anova4_ss_wrapper(means = c(5, 10, 12, 15), sds = c(10, 10, 10, 10), alphalevel = 0.05,
                  powerlevel = 0.8, groupweights = c(.3, .2, .2, .3),
                  Nmin = 60, Nmax = 120, Nchange = 20)
}
