% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossover_power.R
\name{crossover_power}
\alias{crossover_power}
\title{Cross Over Power Calculation}
\usage{
crossover_power(design = "RCS", alpha, J, m, rho, rho_b, xi = 0, d)
}
\arguments{
\item{design}{Either "RCS" for repeated cross-sectional design or "Cohort" for closed cohort design. Default is "RCS".}

\item{alpha}{The significance level or type 1 error rate}

\item{J}{Number of clusters}

\item{m}{The number of individuals in each cluster at each time period}

\item{rho}{Intraclass correlation coefficient (ICC)}

\item{rho_b}{Between period ICC (correlation between outcomes of two individuals in the same cluster but different time periods)}

\item{xi}{Within-cluster, within-subject correlation (correlation between two measurements within the same subject). Note this is different from subject autocorrelation. We expect xi to be larger than rho and rho_b. This is 0 for RCS cross over designs.}

\item{d}{The standardized effect size}
}
\value{
power
}
\description{
Calculate power for either repeated cross sectional or closed cohort cross over designs.
}
\examples{
# Repeated cross sectional cross over power calculation as in Example 12.3
crossover_power(design = "RCS", alpha = 0.05, J = 8, m = 30, rho = 0.05,
                rho_b = 0.025, xi = 0, d = 0.3)
# Closed cohort cross over power calculation as i Example 12.3
crossover_power(design = "Cohort", alpha = 0.05, J = 8, m = 30, rho = 0.05,
                rho_b = 0.025, xi = .4, d = 0.3)
}
