% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossover_ss.R
\name{crossover_ss}
\alias{crossover_ss}
\title{Cross Over Sample Size Calculation}
\usage{
crossover_ss(design = "RCS", alpha, power, m, rho, rho_b, xi = 0, d)
}
\arguments{
\item{design}{Either "RCS" for repeated cross-sectional design or "Cohort" for closed cohort design. Default is "RCS".}

\item{alpha}{The significance level or type 1 error rate}

\item{power}{Specified power level to achieve}

\item{m}{The number of individuals in each cluster at each time period}

\item{rho}{Intraclass correlation coefficient (ICC)}

\item{rho_b}{Between period ICC (correlation between outcomes of two individuals in the same cluster but different time periods)}

\item{xi}{Within-cluster, within-subject correlation (correlation between two measurements within the same subject). Note this is different from subject autocorrelation. We expect xi to be larger than rho and rho_b. This is 0 for RCS cross over designs.}

\item{d}{The standardized effect size}
}
\value{
Calculated number of clusters required and adjusted number of clusters to account for normal approximation
}
\description{
Calculate sample size requirements for either repeated cross sectional or closed cohort cross over designs.
}
\examples{
# repeated cross sectional sample size as in Example 12.4
crossover_ss(design = "RCS", alpha = 0.05, power = 0.8, m = 30, rho = 0.05,
             rho_b = 0.025, xi = 0, d = 0.3)
# closed cohort sample size requirement
crossover_ss(design = "Cohort", alpha = 0.05, power = 0.8, m = 30,
             rho = 0.05, rho_b = 0.025, xi = 0.4, d = 0.3)
}
