% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/es.d.R
\name{es.d}
\alias{es.d}
\title{Cohen's d effect size calculation for one or two means}
\usage{
es.d(delta = NULL, sd = 1)
}
\arguments{
\item{delta}{If one mean: muA (the true mean) - mu0 (the mean under the null). If two means: DeltaA (the true difference mu1 - mu2) - Delta0 (the difference under the null).}

\item{sd}{The estimated standard deviation; defaults to 1.}
}
\value{
A list of the arguments and the d effect size.
}
\description{
Cohen's d effect size calculation for one or two means
}
\examples{
es.d(delta = 6.5 - 5.7, sd = 0.4)
}
