% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/es.fsq.R
\name{es.fsq}
\alias{es.fsq}
\title{Cohen's f^2 effect size calculation for overall F test}
\usage{
es.fsq(Rsq = 0.02)
}
\arguments{
\item{Rsq}{The squared sample multiple correlation coefficient.}
}
\value{
A list of the arguments and the f^2 effect size.
}
\description{
Cohen's f^2 effect size calculation for overall F test
}
\examples{
es.fsq(Rsq = 0.02)
}
