% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/es.q.R
\name{es.q}
\alias{es.q}
\title{Cohen's q effect size calculation for two correlation coefficients}
\usage{
es.q(rho1 = NULL, rho2 = NULL)
}
\arguments{
\item{rho1}{The correlation coefficient in group 1.}

\item{rho2}{The correlation coefficient in group 2.}
}
\value{
A list of the arguments and the q effect size.
}
\description{
Cohen's q effect size calculation for two correlation coefficients
}
\examples{
es.q(rho1 = 0.3, rho2 = 0.1)
}
