% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maps.R
\docType{data}
\name{maps}
\alias{maps}
\title{maps dataframe}
\format{
A data frame with 68 rows and 38 variables:
\describe{
\item{id}{Participant id}
\item{pcs1}{SF-12 Health Survey Physical Component Score at visit 1 (higher means better)}
\item{mcs1}{SF-12 Health Survey Mental Component Score at visit 1 (higher means better)}
\item{psqi1}{Pittsburgh Sleep Quality Index global score at visit 1 (higher means worse)}
\item{cesd1}{Center for Epidemiologic Studies Depression Scale at visit 1 (higher means worse)}
\item{cesdhi1}{Indicator of CES-D score of 16 or higher, indicating risk for clinical depression (1 = yes, 0 = no)}
\item{pss1}{Perceived Stress Scale at visit 1 (higher means worse)}
\item{ies1}{Impact of Events Scale at visit 1 (higher means worse)}
\item{pcs2}{SF-12 PCS at visit 2}
\item{mcs2}{SF-12 MCS at visit 2}
\item{psqi2}{PSQI at visit 2}
\item{cesd2}{CES-D at visit 2}
\item{cesdhi2}{Indicator of CES-D score of 16 or higher at visit 2}
\item{pss2}{PSS at visit 2}
\item{ies2}{IES at visit 2}
\item{pcs3}{SF-12 PCS at visit 3}
\item{mcs3}{SF-12 MCS at visit 3}
\item{psqi3}{PSQI at visit 3}
\item{cesd3}{CES-D at visit 3}
\item{cesdhi3}{Indicator of CES-D score of 16 or higher at visit 3}
\item{pss3}{PSS at visit 3}
\item{ies3}{IES at visit 3}
\item{chemo}{Received chemotherapy for breast cancer (1 = yes, 0 = no)}
\item{radiation}{Received radiation therapy for breast cancer (1 = yes, 0 = no)}
\item{interv_group}{Indicator of intervention group (1 = intervention, 0 = control)}
\item{age_bl}{Age (years) at baseline (visit 1)}
\item{rel_status}{Current relationship status (1 = partnered, 0 = not partnered)}
\item{surgtype}{Type of breast cancer surgery (1 = mastectomy, 0 = lumpectomy)}
\item{educ}{Level of education attained (1 = < college, 2 = college, 3 = >college)}

}
}
\usage{
maps
}
\description{
Data from a subset of the participants in the study described in Bower JE et al. (2015),
Mindfulness meditation for younger breast cancer survivors: A randomized controlled trial.
Cancer, 121: 1231-1240. https://doi.org/10.1002/cncr.29194. The individuals are women
diagnosed with breast cancer at or before age 50. Participants completed assessments at
visits 1 (baseline), 2 and 3.
}
\keyword{datasets}
