% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{margin.sign}
\alias{margin.sign}
\title{Margin Sign Table}
\format{
## `margin.sign`
A data frame with 3 rows and 3 columns:
\describe{
  \item{columns}{Whether the test is for noninferiority of superirority by a margin}
  \item{rows}{Whether a higher or lower parameter value is better}
  \item{xij}{The sign of delta given the row & column}
}
}
\usage{
margin.sign
}
\description{
A table documenting the sign of margin for tests comparing two parameters.
}
\keyword{datasets}
