% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcnemar.test.R
\name{mcnemar.test}
\alias{mcnemar.test}
\title{Power approximation for McNemar's test for two correlated proportions}
\usage{
mcnemar.test(
  N = NULL,
  p1 = NULL,
  p2 = NULL,
  phi = NULL,
  paid = NULL,
  dpr = NULL,
  alpha = 0.05,
  power = NULL,
  sides = 2,
  v = TRUE
)
}
\arguments{
\item{N}{The sample size; the number of pairs.}

\item{p1}{The proportion in condition 1.}

\item{p2}{The proportion in condition 2.}

\item{phi}{The estimated correlation between the two conditions.}

\item{paid}{The smaller of the two discordant probabilities. Either p1, p2, and phi, OR paid and dpr must be specified.}

\item{dpr}{The discordant proportion ratio. Either p1, p2, and phi, OR paid and dpr must be specified.}

\item{alpha}{The significance level or type 1 error rate; defaults to 0.05.}

\item{power}{The specified level of power.}

\item{sides}{Either 1 or 2 (default) to specify a one- or two- sided hypothesis test.}

\item{v}{Either TRUE for verbose output or FALSE to output computed argument only.}
}
\value{
A list of the arguments (including the computed one).
}
\description{
Power approximation for McNemar's test for two correlated proportions
}
\examples{
mcnemar.test(N = NULL, p1 = 0.8, p2 = 0.9, phi = 0, power = 0.9, sides = 2)
mcnemar.test(N = NULL, paid = 0.08, dpr = 0.18 / 0.08, power = 0.9, sides = 2)
}
