% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/McNemar_N.R
\name{McNemar_N}
\alias{McNemar_N}
\title{Sample Size Calculation for McNemar's Test for paired proportions}
\usage{
McNemar_N(
  p1 = NULL,
  p2 = NULL,
  rho = NULL,
  p01 = NULL,
  p10 = NULL,
  alpha = 0.05,
  power = 0.8,
  one_or_twosides = "two"
)
}
\arguments{
\item{p1}{Marginal probability of success for outcome 1. Either NULL or supplied alongside p2 and rho.}

\item{p2}{Marginal probability of success for outcome 2. Either NULL or supplied alongside p1 and rho.}

\item{rho}{Correlation between proportions. Either NULL or supplied alongside p1 and p2.}

\item{p01}{First discordant cell probability. Either NULL or supplied alongside p10.}

\item{p10}{Second discordant cell probability. Either NULL or supplied alongside p01.}

\item{alpha}{Significance level. Default is 0.05}

\item{power}{Required power. Default is 0.8.}

\item{one_or_twosides}{Whether the hypothesis test of interest is one or two-sided. Default is "two".}
}
\value{
Returns a data frame with total number of observations needed (N_obs) and number of pairs needed (N_pairs).
}
\description{
Calculates number of pairs needed to achieve specified level of power from marginal probabilities
and the correlation between probabilities (p1, p2, rho) OR with discordant cell probabilities (p01 and p10).
}
\examples{
# supplying marginal probabilities
McNemar_N(p1 = 0.8, p2 = 0.9, rho = 0, alpha = 0.05, power = 0.9)
# supplying cell probabilities
McNemar_N(p10 = 0.18, p01 = 0.08, alpha = 0.05, power = 0.9)
}
