% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meandiff_ci_power.R
\name{meandiff_ci_power}
\alias{meandiff_ci_power}
\title{Calculate Power of a confidence interval for a Difference of Two Means}
\usage{
meandiff_ci_power(n1, n2, alpha, d = NULL, halfwidth = NULL, sigma = NULL)
}
\arguments{
\item{n1}{The sample size in group 1}

\item{n2}{The sample size in group 2}

\item{alpha}{The significance level}

\item{d}{The standardized halfwidth (halfwidth / sigma). Either d OR halfwidth and sigma need specified.}

\item{halfwidth}{The halfwidth; half of the confidence interval width. If halfwidth is specified, sigma needs specified too.}

\item{sigma}{The estimated standard deviation. If sigma is specified, halfwidth needs specified too.}
}
\value{
The total N, group sample sizes, and corresponding power.
}
\description{
This function calculates the power of a confidence interval for the difference of two means accounting for estimating sigma and allowing for unequal allocation between the two groups.
}
\examples{
# Calculate power for a 95\% confidence interval to have 0.25 standard deviation width
# when there are 100 and 160 people in each group
meandiff_ci_power(n1 = 100, n2 = 160, alpha = 0.05, d = 0.25, halfwidth = NULL, sigma = NULL)
}
\keyword{internal}
