% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meandiff_ci_ss.R
\name{meandiff_ci_ss}
\alias{meandiff_ci_ss}
\title{Calculate Sample Size needed for adequate power for Confidence Interval of Difference in Means.}
\usage{
meandiff_ci_ss(
  N_min,
  N_max,
  ratios = c(1, 1),
  alpha,
  power,
  d = NULL,
  halfwidth = NULL,
  sigma = NULL
)
}
\arguments{
\item{N_min}{The minimum total sample size.}

\item{N_max}{The maximum total sample size}

\item{ratios}{A 2 element vector of group allocation ratios. Equal allocation is default, which is specified as c(1,1). A allocation ratio of 2 (r = n2/n1) would be specified as c(1,2).}

\item{alpha}{The significance level}

\item{power}{The specified power to achieve}

\item{d}{The standardized halfwidth (halfwidth / sigma). Either d OR halfwidth and sigma need specified.}

\item{halfwidth}{The halfwidth; half of the confidence interval width. If halfwidth is specified, sigma needs specified too.}

\item{sigma}{The estimated standard deviation. If sigma is specified, halfwidth needs specified too.}
}
\value{
returns all sample sizes and power that satisfy the specified power level.
}
\description{
Calculate sample sizes in two groups (allowing unequal allocation) that will adequately power the confidence interval of the difference of group means. This function accounts for estimating sigma, the standard deviation.
}
\examples{
# Find sample sizes with allocation ratio of 2 between two groups for a 95\% confidence
# interval to have 80\% probability of a 0.25 standard deviation halfwidth or smaller.
meandiff_ci_ss(N_min = 200, N_max = 325, r = c(1,2),
               alpha = 0.05, power = 0.8, d = 0.25)
}
\keyword{internal}
