% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms.varexplore.R
\name{ms.varexplore}
\alias{ms.varexplore}
\title{Variance exploration for multisite trials with binary outcomes}
\usage{
ms.varexplore(pc, pt)
}
\arguments{
\item{pc}{The probability of the outcome in control clusters.}

\item{pt}{The probability of the outcome in treatment clusters.}
}
\value{
A list of the arguments and a dataframe of outputs.
}
\description{
Variance exploration for multisite trials with binary outcomes
}
\examples{
ms.varexplore(pc = 0.1, pt = 0.2)
}
