% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multendpoints_Ck_ss.R
\name{multendpoints_Ck_ss}
\alias{multendpoints_Ck_ss}
\title{Algorithm and sample size for multiple co-primary endpoint clinical trials}
\usage{
multendpoints_Ck_ss(K, alpha, power, rho, gamma, a, r, delta)
}
\arguments{
\item{K}{The number of K co-primary continuous endpoints. K >= 2.}

\item{alpha}{The significance level.}

\item{power}{The power level.}

\item{rho}{The correlation matrix (K x K) that describes the relationship between each endpoint. Diagonal entries are 1.}

\item{gamma}{The ratio of effect size(s). This can be either a value (K = 2) or a vector (K > 2).}

\item{a}{A vector of length K. For continuous endpoints, a_1 = ... = a_K = 1. For binary endpoints using the chi-square method (without CC), this is a vector where each element is the ratio between sigma_k^star and sigma_k.}

\item{r}{The allocation ratio between the number of people in the control and treatment condition (n_C / n_T).}

\item{delta}{The standardized effect size (value or vector) for each endpoint.}
}
\value{
A dataframe with the solution to the Ck algorithm and the sample size requirements per group.
}
\description{
Calculate the solution to Ck, the integral equation to calculate power for multiple endpoints introduced by Sozu et al in Sample Size Determination in Clinical Trials with Multiple Endpoints 2015.
}
\details{
The sample size provided is the solution to equation (4.5) in the Sozu text. This is the simple formula provided for practical use to calculate sample size requirements as if there were a single endpoint.
Specifically, this formula is n = (Ck + z_alpha)^2 / (kappa * delta_K^2)
}
\examples{
# Following Sozu et al Example in Appendix D.1 with 2 continuous endpoints with equal allocation
multendpoints_Ck_ss(K = 2, alpha = 0.025, power = 0.8,
                    rho = matrix(c(1, 0.5,
                                   0.5, 1), ncol = 2),
                    gamma = 8/7, a = c(1, 1), r = 1, delta = c(0.4, 0.35))
# Following Sozu et al Example in Appendix D.1 with 3 continuous endpoints and equal allocation
delta_vector <- c(0.5, 0.45, 0.4)
gamma_vector <- c(delta_vector[1]/delta_vector[3], delta_vector[2]/delta_vector[3])
rho_matrix <- matrix(c(1, 0.8, 0.8,
                       0.8, 1, 0.5,
                       0.8, 0.5, 1), ncol = 3)

multendpoints_Ck_ss(K = 3, alpha = 0.025, power = 0.8, rho = rho_matrix,
                    gamma = gamma_vector, a = c(1, 1, 1), r = 1,
                    delta = delta_vector)
}
