% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{multisite.data}
\alias{multisite.data}
\title{Multisite Dataframe}
\format{
## `multisite.data`
A data frame with 200 rows and 9 columns:
\describe{
  \item{id}{Individual observation number which goes from id = 1 to 200 (N)}
  \item{j}{Site number which runs from j = 1 to 10 (J)}
  \item{i}{Within site observation number. This runs from i = 1 to 20 (m)}
  \item{xij}{Group level indicator variable coded so that within each site xij sums to 0.}
  \item{Yij}{Outcome variable with a site by group interaction.}
  \item{Wij}{Outcome variable without a site by group interaction.}
}
}
\usage{
multisite.data
}
\description{
A simulated dataset for multisite trials with and without interactions.
}
\keyword{datasets}
