% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multisite.hte.R
\name{multisite.hte}
\alias{multisite.hte}
\title{Power for test of heterogeneity of treatment effect in multisite trials}
\usage{
multisite.hte(
  m = NULL,
  alloc.ratio = 1,
  J = NULL,
  VR = NULL,
  alpha = 0.05,
  v = TRUE
)
}
\arguments{
\item{m}{The total number of subjects per site.}

\item{alloc.ratio}{The allocation ratio of intervention/control per site; defaults to 1.}

\item{J}{The number of sites.}

\item{VR}{The variance ratio (site-level treatment effects / observations within sites) the under the alternative.}

\item{alpha}{The significance level or type 1 error rate; defaults to 0.05.}

\item{v}{Either TRUE for verbose output or FALSE to output computed argument only.}
}
\value{
A list of the arguments (including the computed power).
}
\description{
Power for test of heterogeneity of treatment effect in multisite trials
}
\examples{
multisite.hte(m = 10, J = 30, VR = 8 / 36)
}
