% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newfunction.R
\name{newfunction}
\alias{newfunction}
\title{This is test function}
\usage{
newfunction(a, b)
}
\arguments{
\item{a}{First value to be added}

\item{b}{Second value to be added}
}
\value{
The sum of a + b
}
\description{
This should be new description
}
\examples{
# Try to show new function
newfunction(a=3, b = 6)
}
\keyword{internal}
