% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onemean_ci_power.R
\name{onemean_ci_power}
\alias{onemean_ci_power}
\title{Calculate Power for a Confidence Interval for One Mean}
\usage{
onemean_ci_power(N, d, alpha)
}
\arguments{
\item{N}{The number of observations}

\item{d}{The standardized halfwidth (halfwidth / standard deviation)}

\item{alpha}{The level of significance}
}
\value{
The statistical power associated with the number of observations.
}
\description{
Calculates the power of a confidence interval for a single mean accounting for estimation in sigma, the standard deviation.
}
\examples{
# Calculate probability that a 95\% confidence interval for group size of 73
# to have 0.25 standard deviation halfwidth or less
onemean_ci_power(N = 73, d=0.25, alpha=0.05)
}
\keyword{internal}
