% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onemean_ci_ss.R
\name{onemean_ci_ss}
\alias{onemean_ci_ss}
\title{Sample Size Calculation for One Mean Confidence Interval}
\usage{
onemean_ci_ss(N_min, N_max, d, alpha, power)
}
\arguments{
\item{N_min}{The minimum number of observations to explore power.}

\item{N_max}{The maximum number of observations to explore power.}

\item{d}{The standardized halfwidth (halfwidth / standard deviation).}

\item{alpha}{The significance level.}

\item{power}{The desired level of power.}
}
\value{
The minimum sample size required to achieve the specified level of power.
}
\description{
This function calculates the minimum sample size required to achieve certain precision in estimating a confidence interval for a single mean. This function accounts for estimation in sigma and searches over a range of supplied N.
'Power' used here is the power of a confidence interval, which is not true statistical power; instead it is the probability of the interval obtaining a certain halfwidth.
}
\examples{
# Example 1: Find the required sample size that achieves a half width of 0.25 standard deviations
# at significance level of 0.05 with 80\% power. We specify to search between 62 and 75 people.
onemean_ci_ss(N_min = 62, N_max = 75, d=0.25, alpha=0.05, power=0.8)
}
\keyword{internal}
