% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallelCRT_power.R
\name{parallelCRT_power}
\alias{parallelCRT_power}
\title{Parallel Cluster Randomized Trial Power Calculation}
\usage{
parallelCRT_power(alpha, J, m, d, rho, rho_c, rho_s)
}
\arguments{
\item{alpha}{The significance level or type 1 error rate}

\item{J}{Number of clusters}

\item{m}{The number of individuals in each cluster at each time period}

\item{d}{The standardized effect size}

\item{rho}{Intraclass correlation coefficient (ICC)}

\item{rho_c}{Cluster autocorrelation (proportion of cluster-level variance that is time invariant)}

\item{rho_s}{Subject autocorrelation (This is 0 in repeated cross sectional designs because people are measured only once)}
}
\value{
power
}
\description{
Parallel Cluster Randomized Trial Power Calculation
}
\examples{
# closed cohort design as in Example 12.1
parallelCRT_power(alpha = 0.05, J = 16, m = 30, d = 0.3, rho = 0.05, rho_c = 0.4, rho_s = 0.5)
# repeated cross sectional design as in Example 12.1
parallelCRT_power(alpha = 0.05, J = 16, m = 30, d = 0.3, rho = 0.05, rho_c = 0.4, rho_s = 0)
}
