% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallelCRT_ss.R
\name{parallelCRT_ss}
\alias{parallelCRT_ss}
\title{Parallel Cluster Randomized Trial Sample Size Calculation}
\usage{
parallelCRT_ss(alpha, power, m, d, rho, rho_c, rho_s)
}
\arguments{
\item{alpha}{The significance level or type 1 error rate}

\item{power}{The specified level of power to achieve}

\item{m}{The number of individuals in each cluster at each time period}

\item{d}{The standardized effect size}

\item{rho}{Intraclass correlation coefficient (ICC)}

\item{rho_c}{Cluster autocorrelation (proportion of cluster-level variance that is time invariant)}

\item{rho_s}{Subject autocorrelation (This is 0 in repeated cross sectional designs because people are measured only once)}
}
\value{
Calculated number of clusters required, adjusted number of clusters, and suggested number of clusters that includes the value of K
}
\description{
Parallel Cluster Randomized Trial Sample Size Calculation
}
\examples{
# example 12.2
parallelCRT_ss(alpha = 0.05, power = 0.8, m = 30, d = 0.3, rho = 0.05, rho_c = 0.4, rho_s = 0.5)
}
