% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pss.altprimary.R
\name{pss.altprimary}
\alias{pss.altprimary}
\title{Power calculations for multiple alternative (at least one) primary continuous endpoints assuming known covariance matrix}
\usage{
pss.altprimary(
  K,
  n1 = NULL,
  n.ratio = 1,
  delta = NULL,
  Sigma,
  sd,
  rho,
  alpha = 0.025,
  power = NULL,
  tol = .Machine$double.eps^0.25
)
}
\arguments{
\item{K}{The number of endpoints.}

\item{n1}{The sample size for group 1.}

\item{n.ratio}{The ratio n2/n1 between the sample sizes of two groups; defaults to 1 (equal group sizes).}

\item{delta}{A vector of length K of the true mean differences mu1k - mu2k; must all be positive.}

\item{Sigma}{The covariance matrix of the K outcomes, of dimension K x K.}

\item{sd}{A vector of length K of the standard deviations of the K outcomes.}

\item{rho}{A vector of length 0.5*K*(K-1) of the correlations among the K outcomes.}

\item{alpha}{The significance level or type 1 error rate; defaults to 0.025. A one-sided test is assumed.}

\item{power}{The specified level of power.}
}
\value{
A list of the arguments (including the computed one).
}
\description{
Power calculations for multiple alternative (at least one) primary continuous endpoints assuming known covariance matrix
}
\examples{
pss.altprimary(K = 2, n1 = 100, delta = c(0.4, 0.5), sd = c(1, 1), rho = 0.3,
alpha = 0.025, power = NULL)
}
