% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pss.anova1way.c.bal.R
\name{pss.anova.bal.1w.c}
\alias{pss.anova.bal.1w.c}
\title{Power calculations for one-way balanced analysis of variance contrast test}
\usage{
pss.anova.bal.1w.c(
  n = NULL,
  mvec = NULL,
  cvec = NULL,
  sd = 1,
  Rsq = 0,
  ncov = 0,
  alpha = 0.05,
  power = NULL
)
}
\arguments{
\item{n}{The sample size per group.}

\item{mvec}{A vector of group means c(mu1, mu2, ...).}

\item{cvec}{A vector of contrast coefficients c(c1, c2, ...).}

\item{sd}{The estimated standard deviation within each group; defaults to 1.}

\item{Rsq}{The estimated R^2 for regressing the outcome on the covariates; defaults to 0.}

\item{ncov}{The number of covariates adjusted for in the model; defaults to 0.}

\item{alpha}{The significance level or type 1 error rate; defaults to 0.05.}

\item{power}{The specified level of power.}
}
\value{
A list of the arguments (including the computed one).
}
\description{
Power calculations for one-way balanced analysis of variance contrast test
}
\examples{
pss.anova.bal.1w.c(n = 20, mvec = c(5, 10, 12), cvec = c(1, -1, 0), sd = 10, alpha = 0.025)
pss.anova.bal.1w.c(n = 20, mvec = c(5, 10, 12), cvec = c(1, 0, -1), sd = 10, alpha = 0.025)
}
