% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pss.anova1way.F.bal.R
\name{pss.anova.bal.1w}
\alias{pss.anova.bal.1w}
\title{Power calculations for one-way balanced analysis of variance omnibus F test}
\usage{
pss.anova.bal.1w(
  n = NULL,
  mvec = NULL,
  sd = 1,
  Rsq = 0,
  ncov = 0,
  alpha = 0.05,
  power = NULL
)
}
\arguments{
\item{n}{The sample size per group.}

\item{mvec}{A vector of group means c(mu1, mu2, ...).}

\item{sd}{The estimated standard deviation within each group; defaults to 1.}

\item{Rsq}{The estimated R^2 for regressing the outcome on the covariates; defaults to 0.}

\item{ncov}{The number of covariates adjusted for in the model; defaults to 0.}

\item{alpha}{The significance level or type 1 error rate; defaults to 0.05.}

\item{power}{The specified level of power.}
}
\value{
A list of the arguments (including the computed one).
}
\description{
Power calculations for one-way balanced analysis of variance omnibus F test
}
\examples{
pss.anova.bal.1w(n = 20, mvec = c(5, 10, 12), sd = 10)
pss.anova.bal.1w(n = NULL, mvec = c(-0.25, 0.25), sd = 1, Rsq = 0.5^2, ncov = 1, power = 0.8)
}
