% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pss.anova.bal.2w.R
\name{pss.anova.bal.2w}
\alias{pss.anova.bal.2w}
\title{Power calculations for two-way balanced analysis of variance omnibus F test}
\usage{
pss.anova.bal.2w(
  n = NULL,
  mmatrix = NULL,
  sd = 1,
  Rsq = 0,
  ncov = 0,
  alpha = 0.05,
  power = NULL
)
}
\arguments{
\item{n}{The sample size per group.}

\item{mmatrix}{A matrix of group means (see example).}

\item{sd}{The estimated standard deviation within each group; defaults to 1.}

\item{Rsq}{The estimated R^2 for regressing the outcome on the covariates; defaults to 0.}

\item{ncov}{The number of covariates adjusted for in the model; defaults to 0.}

\item{alpha}{The significance level or type 1 error rate; defaults to 0.05.}

\item{power}{The specified level of power.}
}
\value{
A list of the arguments (including the computed one).
}
\description{
Power calculations for two-way balanced analysis of variance omnibus F test
}
\examples{
# Example 5.8
mmatrix <- matrix(c(9.3, 8.9, 8.5, 8.7, 8.3, 7.9), nrow = 2, byrow = TRUE)
pss.anova.bal.2w(n = 30, mmatrix = mmatrix, sd = 2, alpha = 0.05)
# Example 5.10
mmatrix <- matrix(c(9.3, 8.9, 8.5, 8.7, 8.3, 7.3), nrow = 2, byrow = TRUE)
pss.anova.bal.2w(n = 30, mmatrix = mmatrix, sd = 2, alpha = 0.05)
# Example 5.14
mmatrix <- matrix(c(9.3, 8.9, 8.5, 8.7, 8.3, 7.9), nrow = 2, byrow = TRUE)
pss.anova.bal.2w(n = 30, mmatrix = mmatrix, sd = 2, Rsq = 0.4, ncov = 1, alpha = 0.05)
}
