% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pss.anova.f.es.R
\name{pss.anova.f.es}
\alias{pss.anova.f.es}
\title{Cohen's f effect size calculation for one- or two- way analysis of variance}
\usage{
pss.anova.f.es(means = NULL, sd = NULL)
}
\arguments{
\item{sd}{The estimated standard deviation within each group.}

\item{mvec}{A vector of group means. One of mvec OR mmatrix must be specified.}

\item{mmatrix}{A matrix of group means. One of mvec OR mmatrix must be specified.}
}
\value{
A list of the arguments and various f effect sizes.
}
\description{
Cohen's f effect size calculation for one- or two- way analysis of variance
}
\examples{
pss.anova.f.es(means = c(5, 10, 12), sd = 10)
mmatrix <- matrix(c(9.3, 8.9, 8.5, 8.7, 8.3, 7.9), nrow = 2, byrow = TRUE)
pss.anova.f.es(means = mmatrix, sd = 2)
}
