% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pss.anova.unbal.1w.c.R
\name{pss.anova.unbal.1w.c}
\alias{pss.anova.unbal.1w.c}
\title{Power calculations for one-way unbalanced analysis of variance contrast test}
\usage{
pss.anova.unbal.1w.c(
  nvec = NULL,
  mvec = NULL,
  cvec = NULL,
  sd = NULL,
  Rsq = 0,
  ncov = 0,
  alpha = 0.05
)
}
\arguments{
\item{nvec}{A vector of group sample sizes c(n1, n2, ...).}

\item{mvec}{A vector of group mvec c(mu1, mu2, ...).}

\item{cvec}{A vector of contrast cvecicients c(c1, c2, ...).}

\item{sd}{The estimated standard deviation within each group.}

\item{Rsq}{The estimated R^2 for regressing the outcome on the covariates; defaults to 0.}

\item{ncov}{The number of covariates adjusted for in the model; defaults to 0.}

\item{alpha}{The significance level or type 1 error rate; defaults to 0.05.}
}
\value{
A list of the arguments (including the computed power).
}
\description{
Power calculations for one-way unbalanced analysis of variance contrast test
}
\examples{
# Example 5.5
pss.anova.unbal.1w.c(nvec = c(20, 20, 20), mvec = c(5, 10, 12), cvec = c(1, -1, 0), sd = 10, alpha = 0.025)
pss.anova.unbal.1w.c(nvec = c(20, 20, 20), mvec = c(5, 10, 12), cvec = c(1, 0, -1), sd = 10, alpha = 0.025)
}
