% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pss.anova2way.c.unbal.R
\name{pss.anova.unbal.2w.c}
\alias{pss.anova.unbal.2w.c}
\title{Power calculations for two-way unbalanced analysis of variance contrast test}
\usage{
pss.anova.unbal.2w.c(
  nmatrix = nmatrix,
  mmatrix = NULL,
  cvec = NULL,
  factor = c("a", "b"),
  sd = NULL,
  Rsq = 0,
  ncov = 0,
  alpha = 0.05
)
}
\arguments{
\item{nmatrix}{A matrix of group sample sizes (see example).}

\item{mmatrix}{A matrix of group means (see example).}

\item{cvec}{A vector of contrast coefficients c(c1, c2, ...).}

\item{factor}{Either "a" or "b" depending on which factor the contrast test is being made on.}

\item{sd}{The estimated standard deviation within each group.}

\item{Rsq}{The estimated R^2 for regressing the outcome on the covariates; defaults to 0.}

\item{ncov}{The number of covariates adjusted for in the model; defaults to 0.}

\item{alpha}{The significance level or type 1 error rate; defaults to 0.05.}

\item{power}{The specified level of power.}
}
\value{
A list of the arguments (including the computed one).
}
\description{
Power calculations for two-way unbalanced analysis of variance contrast test
}
\examples{
nmatrix <- matrix(c(30, 30, 30, 30, 30, 30), nrow = 2, byrow = TRUE)
mmatrix <- matrix(c(9.3, 8.9, 8.5, 8.7, 8.3, 7.3), nrow = 2, byrow = TRUE)
pss.anova.unbal.2w.c(nmatrix = nmatrix, mmatrix = mmatrix, cvec = c(1, 0, -1), factor = "b", sd = 2, alpha = 0.05)
}
