% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pss.anova2way.se.unbal.R
\name{pss.anova.unbal.2w.se}
\alias{pss.anova.unbal.2w.se}
\title{Power calculations for two-way unbalanced analysis of variance simple effects test}
\usage{
pss.anova.unbal.2w.se(
  nmatrix = NULL,
  mmatrix = NULL,
  cmatrix = NULL,
  sd = 0,
  Rsq = 0,
  ncov = 0,
  alpha = 0.05
)
}
\arguments{
\item{nmatrix}{A matrix of sample sizes (see example).}

\item{mmatrix}{A matrix of group means (see example).}

\item{cmatrix}{A matrix of contrast coefficients (see example).}

\item{sd}{The estimated standard deviation within each group.}

\item{Rsq}{The estimated R^2 for regressing the outcome on the covariates; defaults to 0.}

\item{ncov}{The number of covariates adjusted for in the model; defaults to 0.}

\item{alpha}{The significance level or type 1 error rate; defaults to 0.05.}
}
\value{
A list of the arguments (including the computed power).
}
\description{
Power calculations for two-way unbalanced analysis of variance simple effects test
}
\examples{
nmatrix <- matrix(c(30, 30, 30, 30, 30, 30), nrow = 2, byrow = TRUE)
mmatrix <- matrix(c(9.3, 8.9, 8.5, 8.7, 8.3, 7.3), nrow = 2, byrow = TRUE)
cmatrix <- matrix(c(-1, 0, 0, 1, 0, 0), nrow = 2, byrow = TRUE)
pss.anova.unbal.2w.se(nmatrix = nmatrix, mmatrix = mmatrix, cmatrix = cmatrix, sd = 2, alpha = 0.025)
}
