% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pss.anova2way.se.bal.R
\name{pss.anova2way.se.bal}
\alias{pss.anova2way.se.bal}
\title{Power calculations for two-way balanced analysis of variance simple effects test}
\usage{
pss.anova2way.se.bal(
  n = NULL,
  mmatrix = NULL,
  cmatrix = NULL,
  sd = 1,
  Rsq = 0,
  ncov = 0,
  alpha = 0.05,
  power = NULL,
  v = TRUE
)
}
\arguments{
\item{n}{The sample size per group.}

\item{mmatrix}{A matrix of group means (see example).}

\item{cmatrix}{A matrix of contrast coefficients (see example).}

\item{sd}{The estimated standard deviation within each group; defaults to 1.}

\item{Rsq}{The estimated R^2 for regressing the outcome on the covariates; defaults to 0.}

\item{ncov}{The number of covariates adjusted for in the model; defaults to 0.}

\item{alpha}{The significance level or type 1 error rate; defaults to 0.05.}

\item{power}{The specified level of power.}

\item{v}{Either TRUE for verbose output or FALSE to output computed argument only.}
}
\value{
A list of the arguments (including the computed one).
}
\description{
Power calculations for two-way balanced analysis of variance simple effects test
}
\examples{
mmatrix <- matrix(c(9.3, 8.9, 8.5, 8.7, 8.3, 7.3), nrow = 2, byrow = TRUE)
cmatrix <- matrix(c(-1, 0, 0, 1, 0, 0), nrow = 2, byrow = TRUE)
pss.anova2way.se.bal(n = 30, mmatrix = mmatrix, cmatrix = cmatrix, sd = 2, alpha = 0.025)
}
