% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pss.chisq.indep.R
\name{pss.chisq.indep}
\alias{pss.chisq.indep}
\title{Power calculations for chi-square test of independence}
\usage{
pss.chisq.indep(pmatrix = NULL, n = NULL, alpha = 0.05, power = NULL)
}
\arguments{
\item{pmatrix}{The two-way probability table under the alternative hypothesis.}

\item{n}{The number of total observations.}

\item{alpha}{The significance level or type 1 error rate; defaults to 0.05.}

\item{power}{The specified level of power.}
}
\value{
A list of the arguments (including the computed one).
}
\description{
Power calculations for chi-square test of independence
}
\examples{
# Example 8.2
pss.chisq.indep(pmatrix = matrix(c(0.050, 0.350, 0.100, 0.075, 0.250, 0.175), nrow = 2, byrow = TRUE), n = 230)
# Example 8.3
pss.chisq.indep(pmatrix = matrix(c(0.3, 0.2, 0.4, 0.1), nrow = 2, byrow = TRUE), n = 200)
}
