% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pss.ci.mean.R
\name{pss.ci.mean}
\alias{pss.ci.mean}
\title{Power calculations for precision analysis for one mean}
\usage{
pss.ci.mean(
  n = NULL,
  h = NULL,
  sd = 1,
  alpha = 0.05,
  power = NULL,
  cond = FALSE
)
}
\arguments{
\item{n}{The sample size.}

\item{h}{The desired halfwidth.}

\item{sd}{The estimated standard deviation; defaults to 1.}

\item{alpha}{The significance level or type 1 error rate; defaults to 0.05.}

\item{power}{The specified level of power.}

\item{cond}{Specify using unconditional or conditional probability. Defaults to FALSE.}
}
\value{
A list of the arguments (including the computed one).
}
\description{
Power calculations for precision analysis for one mean
}
\examples{
# Example 9.2
pss.ci.mean(n = NULL, h = 0.25, power = 0.8)
pss.ci.mean(n = 62, h = 0.25, power = NULL)
# Example 9.3
pss.ci.mean(n = 73, h = 0.25, cond = TRUE)
}
