% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pss.ci.meandiff.R
\name{pss.ci.meandiff}
\alias{pss.ci.meandiff}
\title{Power calculations for precision analysis for a difference between means}
\usage{
pss.ci.meandiff(
  n = NULL,
  n.ratio = 1,
  h = NULL,
  sd = 1,
  alpha = 0.05,
  power = NULL,
  cond = FALSE
)
}
\arguments{
\item{n}{The sample size for group 1.}

\item{n.ratio}{The ratio n2/n1 between the sample sizes of two groups; defaults to 1 (equal group sizes).}

\item{h}{The desired halfwidth for the difference in means.}

\item{sd}{The estimated standard deviation; defaults to 1.}

\item{alpha}{The significance level or type 1 error rate; defaults to 0.05.}

\item{power}{The specified level of power.}

\item{cond}{Specify using unconditional or conditional probability. Defaults to FALSE.}
}
\value{
A list of the arguments (including the computed one).
}
\description{
Power calculations for precision analysis for a difference between means
}
\examples{
pss.ci.meandiff(n = NULL, h = 0.25, power = 0.8)
pss.ci.meandiff(n = 134, h = 0.25, cond = TRUE)
}
