% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pss.corr.1samp.R
\name{pss.corr.1samp}
\alias{pss.corr.1samp}
\title{Power calculations for one correlation coefficient}
\usage{
pss.corr.1samp(n = NULL, rho0 = 0, rhoA = NULL, alpha = 0.05, power = NULL)
}
\arguments{
\item{n}{The sample size.}

\item{rho0}{The correlation coefficient under the null hypothesis; defaults to 0.}

\item{rhoA}{The correlation coefficient under the alternative hypothesis.}

\item{alpha}{The significance level or type 1 error rate; defaults to 0.05.}

\item{power}{The specified level of power.}
}
\value{
A list of the arguments (including the computed one).
}
\description{
Power calculations for one correlation coefficient
}
\examples{
#' # Example 10.1
pss.corr.1samp(n = 100, rhoA = 0.2)
pss.corr.1samp(n = 100, rho0 = 0.2, rhoA = 0.4)
}
