% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pss.corr.2samp.R
\name{pss.corr.2samp}
\alias{pss.corr.2samp}
\title{Power calculations for comparing two correlation coefficients}
\usage{
pss.corr.2samp(
  n = NULL,
  n.ratio = 1,
  rho1 = NULL,
  rho2 = NULL,
  alpha = 0.05,
  power = NULL,
  sides = 2
)
}
\arguments{
\item{n}{The sample size for group 1.}

\item{n.ratio}{The ratio n2/n1 between the sample sizes of two groups; defaults to 1 (equal group sizes).}

\item{rho1}{The correlation coefficient in the first group.}

\item{rho2}{The correlation coefficient in the second group.}

\item{alpha}{The significance level or type 1 error rate; defaults to 0.05.}

\item{power}{The specified level of power.}

\item{sides}{Either 1 or 2 (default) to specify a one- or two- sided hypothesis test.}
}
\value{
A list of the arguments (including the computed one).
}
\description{
Power calculations for comparing two correlation coefficients
}
\examples{
pss.corr.2samp(n = 300, rho1 = 0.3, rho2 = 0.1, sides = 1)
}
