% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pss.crt.parallel.bin.R
\name{pss.crt.parallel.bin}
\alias{pss.crt.parallel.bin}
\title{Power for a cluster randomized trial with a binary outcome}
\usage{
pss.crt.parallel.bin(
  m = NULL,
  m.sd = 0,
  J = NULL,
  pc = NULL,
  pt = NULL,
  sigma.u = NULL,
  alpha = 0.05,
  power = NULL,
  sides = 2
)
}
\arguments{
\item{m}{The number of subjects per cluster.}

\item{m.sd}{The standard deviation of cluster sizes (provide if unequal number of participants per cluster); defaults to 0.}

\item{J}{The number of clusters.}

\item{pc}{The probability of the outcome in control clusters.}

\item{pt}{The probability of the outcome in treatment clusters.}

\item{sigma.u}{Standard deviation of the cluster random effect.}

\item{alpha}{The significance level or type 1 error rate; defaults to 0.05.}

\item{power}{The specified level of power.}

\item{sides}{Either 1 or 2 (default) to specify a one- or two- sided hypothesis test.}
}
\value{
A list of the arguments (including the computed one).
}
\description{
Power for a cluster randomized trial with a binary outcome
}
\examples{
pss.crt.parallel.bin(m = 60, J = NULL, pc = 0.25, pt = 0.15, sigma.u = 0.3, power = 0.8)
pss.crt.parallel.bin(m = 60, m.sd = 1, J = NULL, pc = 0.25, pt = 0.15, sigma.u = 0.3, power = 0.8)
}
