% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pss.crt.varexplore.R
\name{pss.crt.varexplore}
\alias{pss.crt.varexplore}
\title{Variance exploration for cluster randomized trials with binary outcomes}
\usage{
pss.crt.varexplore(pc, pt)
}
\arguments{
\item{pc}{The probability of the outcome in control clusters.}

\item{pt}{The probability of the outcome in treatment clusters.}
}
\value{
A list of the arguments and a dataframe of outputs.
}
\description{
Variance exploration for cluster randomized trials with binary outcomes
}
\examples{
pss.crt.varexplore(pc = 0.25, pt = 0.15)
}
