% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pss.crt.xo.cont.R
\name{pss.crt.xo.cont}
\alias{pss.crt.xo.cont}
\title{Power for test of treatment effect in 2x2 crossover cluster randomized trial}
\usage{
pss.crt.xo.cont(
  m = NULL,
  J.arm = NULL,
  delta = NULL,
  sd = 1,
  icc = 0,
  cac = 1,
  sac = 0,
  alpha = 0.05,
  power = NULL,
  sides = 2
)
}
\arguments{
\item{m}{The number of subjects measured during each cluster-period.}

\item{J.arm}{The number of clusters in each arm.}

\item{delta}{The difference between the intervention and control means under the alternative minus the difference under the null hypothesis.}

\item{sd}{The total standard deviation of the outcome variable; defaults to 1.}

\item{icc}{The within-cluster, within-period intraclass correlation coefficient; defaults to 0.}

\item{cac}{The cluster autocorrelation; defaults to 1.}

\item{sac}{The subject autocorrelation; defaults to 0.}

\item{alpha}{The significance level or type 1 error rate; defaults to 0.05.}

\item{power}{The specified level of power.}

\item{sides}{Either 1 or 2 (default) to specify a one- or two- sided hypothesis test.}

\item{iccb}{The within-cluster, between-period intraclass correlation coefficient; defaults to 0.}

\item{xi}{The  within-cluster, within-subject correlation (correlation between two measurements within the same subject); defaults to 0.}
}
\value{
A list of the arguments (including the computed one).
}
\description{
Power for test of treatment effect in 2x2 crossover cluster randomized trial
}
\examples{
pss.crt.xo.cont(m = 30, J.arm = 4, delta = 0.3, icc = 0.05, cac = 0.8, sac = 0.4)
}
