% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pss.es.anova.f.R
\name{pss.es.anova.f}
\alias{pss.es.anova.f}
\title{Cohen's f effect size calculation for one- or two- way analysis of variance}
\usage{
pss.es.anova.f(means = NULL, sd = NULL)
}
\arguments{
\item{means}{A vector or matrix of group means.}

\item{sd}{The estimated standard deviation within each group.}
}
\value{
A list of the arguments and various f effect sizes.
}
\description{
Cohen's f effect size calculation for one- or two- way analysis of variance
}
\examples{
pss.es.anova.f(means = c(5, 10, 12), sd = 10)
mmatrix <- matrix(c(9.3, 8.9, 8.5, 8.7, 8.3, 7.9), nrow = 2, byrow = TRUE)
pss.es.anova.f(means = mmatrix, sd = 2)
}
