% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pss.es.h.R
\name{pss.es.h}
\alias{pss.es.h}
\title{Cohen's h effect size calculation for two proportions}
\usage{
pss.es.h(p1 = NULL, p2 = NULL)
}
\arguments{
\item{p1}{The proportion in group 1.}

\item{p2}{The proportion in group 2.}
}
\value{
A list of the arguments and the h effect size.
}
\description{
Cohen's h effect size calculation for two proportions
}
\examples{
pss.es.h(p1 = 0.8, p2 = 0.6)
}
