% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pss.es.w.R
\name{pss.es.w}
\alias{pss.es.w}
\title{Cohen's w effect size calculation for chi-square tests}
\usage{
pss.es.w(p0vec = NULL, p1vec = NULL, pmatrix = NULL)
}
\arguments{
\item{p0vec}{The first vector of probabilities. Both p0vec and p1vec, or pmatrix must be specified.}

\item{p1vec}{The second vector of probabilities. Both p0vec and p1vec, or pmatrix must be specified.}

\item{pmatrix}{The two-way probability table. Both p0vec and p1vec, or pmatrix must be specified.}
}
\value{
A list of the arguments and the w effect size.
}
\description{
Cohen's w effect size calculation for chi-square tests
}
\examples{
pss.es.w(p0vec = c(0.5, 0.3, 0.2), p1vec = c(0.7, 0.2, 0.1))
pss.es.w(pmatrix = matrix(c(0.050, 0.350, 0.100, 0.075, 0.250, 0.175), nrow = 2, byrow = TRUE))
}
