% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pss.irgtt.bin.R
\name{pss.irgtt.bin}
\alias{pss.irgtt.bin}
\title{Power for an individual randomized group treatment trial with binary outcomes}
\usage{
pss.irgtt.bin(
  m = NULL,
  J = NULL,
  n = NULL,
  p1 = NULL,
  p2 = NULL,
  icc = 0,
  alpha = 0.05,
  power = NULL,
  sides = 2,
  v = TRUE
)
}
\arguments{
\item{m}{The number of subjects per cluster in the treatment group.}

\item{J}{The number of clusters in the treatment group.}

\item{n}{The number of total participants in the control group.}

\item{p1}{The probability of the outcome in the control group.}

\item{p2}{The probability of the outcome in the intervention group.}

\item{icc}{The intraclass correlation coefficient in the treatment group; defaults to 0.}

\item{alpha}{The significance level or type 1 error rate; defaults to 0.05.}

\item{power}{The specified level of power.}

\item{sides}{Either 1 or 2 (default) to specify a one- or two- sided hypothesis test.}

\item{v}{Either TRUE for verbose output or FALSE to output computed argument only.}
}
\value{
A list of the arguments (including the computed one).
}
\description{
Power for an individual randomized group treatment trial with binary outcomes
}
\examples{
pss.irgtt.bin(m = 20, J = 5, n = 100, p1 = 0.8, p2 = 0.6, icc = 0.04, sides = 2)
pss.irgtt.bin(m = 20, J = 6, n = 120, p1 = 0.8, p2 = 0.6, icc = 0.04, sides = 2)
}
