% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pss.irgtt.cont.R
\name{pss.irgtt.cont}
\alias{pss.irgtt.cont}
\title{Power for individually randomized group treatment trial with continuous outcome}
\usage{
pss.irgtt.cont(
  m = NULL,
  J = NULL,
  n = NULL,
  delta = NULL,
  sd = 1,
  icc = 0,
  Theta = 1,
  alpha = 0.05,
  power = NULL,
  sides = 2,
  tol = .Machine$double.eps^0.25,
  v = TRUE
)
}
\arguments{
\item{m}{The number of subjects per cluster in the treatment group.}

\item{J}{The number of clusters in the treatment group.}

\item{n}{The number of total participants in the control group.}

\item{delta}{The difference between the intervention and control means under the alternative minus the difference under the null hypothesis.}

\item{sd}{The total standard deviation of the outcome variable in the control group; defaults to 1.}

\item{icc}{The intraclass correlation coefficient in the treatment group; defaults to 0.}

\item{Theta}{The ratio of the total variance between intervention and control groups; defaults to 1.}

\item{alpha}{The significance level or type 1 error rate; defaults to 0.05.}

\item{power}{The specified level of power.}

\item{sides}{Either 1 or 2 (default) to specify a one- or two- sided hypothesis test.}

\item{v}{Either TRUE for verbose output or FALSE to output computed argument only.}
}
\value{
A list of the arguments (including the computed one).
}
\description{
Power for individually randomized group treatment trial with continuous outcome
}
\examples{
pss.irgtt.cont(m = 10, J = 12, n = 120, delta = 0.4, icc = 0.05, Theta = 1, power = NULL)
pss.irgtt.cont(m = 10, J = 12, n = NULL, delta = 0.4, icc = 0.05, Theta = 1, power = 0.8)
}
