% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pss.mlrF.overall.R
\name{pss.mlrF.overall}
\alias{pss.mlrF.overall}
\title{Power calculations for a multiple linear regression overall F test}
\usage{
pss.mlrF.overall(
  n = NULL,
  p = NULL,
  Rsq = NULL,
  fsq = NULL,
  alpha = 0.05,
  power = NULL,
  random = FALSE
)
}
\arguments{
\item{n}{The sample size.}

\item{p}{The number of predictors.}

\item{Rsq}{The squared sample multiple correlation coefficient.}

\item{fsq}{The squared f effect size. Either Rsq OR fsq must be specified.}

\item{alpha}{The significance level or type 1 error rate; defaults to 0.05.}

\item{power}{The specified level of power.}

\item{random}{Whether the values of the predictors are random; defaults to FALSE.}
}
\value{
A list of the arguments (including the computed one).
}
\description{
Power calculations for a multiple linear regression overall F test
}
\examples{
# Example 10.6
pss.mlrF.overall(n = 400, p = 2, Rsq = 0.02)
pss.mlrF.overall(n = 400, p = 2, fsq = 0.02 / (1 - 0.02))
# Example 10.7
pss.mlrF.overall(n = 109, p = 1, Rsq = 0.3^2)
# Example 10.8
pss.mlrF.overall(n = 50, p = 1, Rsq = 0.2)
pss.mlrF.overall(n = 50, p = 3, Rsq = 0.2)
pss.mlrF.overall(n = 50, p = 5, Rsq = 0.2)
# Example 10.9
pss.mlrF.overall(n = 400, p = 2, Rsq = 0.02, random = TRUE)
}
