% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pss.multi.ate.R
\name{pss.multi.ate}
\alias{pss.multi.ate}
\title{Power for test of average treatment effect}
\usage{
pss.multi.ate(m, J, r = 1, gamma, rho0, rho1, ssq.Y, alpha = 0.05)
}
\arguments{
\item{m}{The number of subjects per site.}

\item{J}{The number of sites.}

\item{r}{The allocation ratio to intervention and control at each site; defaults to 1.}

\item{gamma}{The average treatment effect under the alternative.}

\item{rho0}{The proportion of total variance of the outcome attributable to variation in site-level means.}

\item{rho1}{The proportion of total variance of the outcome attributable to variation in the treatment effect across sites.}

\item{ssq.Y}{The total variance of the outcome variable Y.}

\item{alpha}{The significance level or type 1 error rate; defaults to 0.05.}
}
\value{
A list of the arguments (including the computed power).
}
\description{
Power for test of average treatment effect
}
\examples{
# Example 14.2
pss.multi.ate(m = 20, J = 10, gamma = 3, rho0 = 0.1, rho1 = 0, ssq.Y = 40)
pss.multi.ate(m = 20, J = 10, gamma = 3, rho0 = 0.095, rho1 = 0.048, ssq.Y = 48)
# Example 14.3
pss.multi.ate(m = 20, J = 10, r = 1.5, gamma = 3, rho0 = 0.095, rho1 = 0.048, ssq.Y = 48)
}
