% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pss.multi.hte.R
\name{pss.multi.hte}
\alias{pss.multi.hte}
\title{Power for test of heterogeneity of treatment effect}
\usage{
pss.multi.hte(m, J, ssq.u1, ssq.e, alpha = 0.05)
}
\arguments{
\item{m}{The number of subjects per site.}

\item{J}{The number of sites.}

\item{ssq.u1}{The variance of the site-level treatment effects (sigma squared u1) under the alternative.}

\item{ssq.e}{The variance of the observations within sites (sigma squared epsilon) under the alternative.}

\item{alpha}{The significance level or type 1 error rate; defaults to 0.05.}
}
\value{
A list of the arguments (including the computed power).
}
\description{
Power for test of heterogeneity of treatment effect
}
\examples{
# Example 14.7
pss.multi.hte(m = 10, J = 30, ssq.u1 = 8, ssq.e = 36)
}
