% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pss.multisite.ate.unbal.R
\name{pss.multisite.ate.unbal}
\alias{pss.multisite.ate.unbal}
\title{Power for test of average treatment effect}
\usage{
pss.multisite.ate.unbal(
  m = NULL,
  m.ratio = NULL,
  J = NULL,
  gamma = NULL,
  rho0 = NULL,
  rho1 = NULL,
  ssq.Y = NULL,
  d = NULL,
  alpha = 0.05,
  sides = 2
)
}
\arguments{
\item{m}{The number of subjects in the control for each site.}

\item{m.ratio}{The allocation ratio per site}

\item{J}{The number of sites.}

\item{gamma}{The average treatment effect under the alternative.}

\item{rho0}{The proportion of total variance of the outcome attributable to variation in site-level means.}

\item{rho1}{The proportion of total variance of the outcome attributable to variation in the treatment effect across sites.}

\item{ssq.Y}{The total variance of the outcome variable Y.}

\item{alpha}{The significance level or type 1 error rate; defaults to 0.05.}

\item{sides}{Either 1 or 2 (default) to specify a one- or two- sided hypothesis test.}
}
\value{
A list of the arguments (including the computed power).
}
\description{
Power for test of average treatment effect
}
\examples{
pss.multisite.ate.unbal(m = 20, m.ratio = 1.5, J = 10, gamma = 3, rho0 = 0.0952381, rho1 = 0.04761905, ssq.Y = 48)
}
\keyword{internal}
