% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pss.multisite.hte.R
\name{pss.multisite.hte}
\alias{pss.multisite.hte}
\title{Power for test of heterogeneity of treatment effect in multisite trials}
\usage{
pss.multisite.hte(m = NULL, alloc.ratio = 1, J = NULL, VR = NULL, alpha = 0.05)
}
\arguments{
\item{m}{The total number of subjects per site.}

\item{J}{The number of sites.}

\item{VR}{The variance ratio (site-level treatment effects / observations within sites) the under the alternative.}

\item{alpha}{The significance level or type 1 error rate; defaults to 0.05.}
}
\value{
A list of the arguments (including the computed power).
}
\description{
Power for test of heterogeneity of treatment effect in multisite trials
}
\examples{
pss.multisite.hte(m = 10, J = 30, VR = 8 / 36)
}
