% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pss.multisite.re.R
\name{pss.multisite.re}
\alias{pss.multisite.re}
\title{Relative efficiency due to unequal number of participants per site}
\usage{
pss.multisite.re(m.mean = NULL, m.sd = NULL, rho = NULL)
}
\arguments{
\item{m.mean}{The mean cluster size.}

\item{m.sd}{The standard deviation of cluster sizes.}

\item{rho}{The intraclass correlation among cluster sizes.}
}
\value{
A list of the arguments (including the computed relative efficiency).
}
\description{
Relative efficiency due to unequal number of participants per site
}
\examples{
1 / pss.multisite.re(m.mean = 30, m.sd = 23, rho = 0.05)$re
1 / pss.multisite.re(m.mean = 30, m.sd = 23, rho = 0.1)$re

}
