% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pss.prop.1samp.R
\name{pss.prop.1samp}
\alias{pss.prop.1samp}
\title{Power calculations for one sample proportion tests}
\usage{
pss.prop.1samp(
  N = NULL,
  p0 = NULL,
  pA = NULL,
  alpha = 0.05,
  power = NULL,
  sides = 2
)
}
\arguments{
\item{N}{The sample size.}

\item{p0}{The proportion under the null.}

\item{pA}{The true proportion.}

\item{alpha}{The significance level or type 1 error rate; defaults to 0.05.}

\item{power}{The specified level of power.}

\item{sides}{Either 1 or 2 (default) to specify a one- or two- sided hypothesis test.}
}
\value{
A list of the arguments (including the computed one).
}
\description{
Power calculations for one sample proportion tests
}
\examples{
pss.prop.1samp(N = NULL, p0 = 0.2, pA = 0.3, power = 0.8, sides = 1)
pss.prop.1samp(N = NULL, p0 = 0.4, pA = 0.5, power = 0.8, sides = 1)

}
