% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pss.prop.2samp.R
\name{pss.prop.2samp}
\alias{pss.prop.2samp}
\title{Power calculations for two sample proportion tests}
\usage{
pss.prop.2samp(
  n = NULL,
  p1 = NULL,
  p2 = NULL,
  delta = 0,
  alpha = 0.05,
  power = NULL,
  n.ratio = 1,
  sides = 2
)
}
\arguments{
\item{n}{The sample size for group 1.}

\item{p1}{The proportion in group 1.}

\item{p2}{The proportion in group 2.}

\item{delta}{The margin of noninferiority or superiority; defaults to 0. See delta.sign for guidance on the sign of delta.}

\item{alpha}{The significance level or type 1 error rate; defaults to 0.05.}

\item{power}{The specified level of power.}

\item{n.ratio}{The ratio n2/n1 between the sample sizes of two groups; defaults to 1 (equal group sizes).}

\item{sides}{Either 1 or 2 (default) to specify a one- or two- sided hypothesis test.}
}
\value{
A list of the arguments (including the computed one).
}
\description{
Power calculations for two sample proportion tests
}
\examples{
pss.prop.2samp(n = NULL, p1 = 0.6, p2 = 0.8, alpha = 0.025, power = 0.9, sides = 1)
pss.prop.2samp(n = NULL, p1 = 0.25, p2 = 0.25, delta = 0.1, alpha = 0.025, power = 0.8, sides = 1)
}
