% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pss.prop.test.equiv.R
\name{pss.prop.test.equiv}
\alias{pss.prop.test.equiv}
\title{Power calculations for test of equivalence of two proportions}
\usage{
pss.prop.test.equiv(
  n1 = NULL,
  n.ratio = 1,
  p1 = NULL,
  p2 = NULL,
  margin = NULL,
  alpha = 0.05,
  power = NULL,
  sides = 2
)
}
\arguments{
\item{n1}{The sample size for group 1.}

\item{n.ratio}{The ratio n2/n1 between the sample sizes of two groups; defaults to 1 (equal group sizes).}

\item{p1}{The outcome proportion in group 1.}

\item{p2}{The outcome proportion in group 2.}

\item{margin}{The equivalence margin. See margin.sign for guidance on the sign of margin.}

\item{alpha}{The significance level or type 1 error rate; defaults to 0.05.}

\item{power}{The specified level of power.}

\item{sides}{Either 1 or 2 (default) to specify a one- or two- sided hypothesis test.}
}
\value{
A list of the arguments (including the computed one).
}
\description{
Power calculations for test of equivalence of two proportions
}
\examples{
pss.prop.test.equiv(n1 = NULL, p1 = 0.5, p2 = 0.5, margin = 0.1, alpha = 0.05, power = 0.8,
sides = 1)
}
