% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pss.re.crt.bin.R
\name{pss.re.crt.bin}
\alias{pss.re.crt.bin}
\title{Relative efficiency for cluster randomized trials with binary outcomes
Relative efficiency of varying to equal cluster sizes}
\usage{
pss.re.crt.bin(m, m.sd, pc, pt, sigma.u)
}
\arguments{
\item{m}{The number of subjects per cluster or the mean cluster size (if unequal number of participants per cluster).}

\item{m.sd}{The standard deviation of cluster sizes (provide if unequal number of participants per cluster).}

\item{pc}{The probability of the outcome in control clusters.}

\item{pt}{The probability of the outcome in treatment clusters.}

\item{sigma.u}{Standard deviation of the cluster random effect.}
}
\value{
The computed RE.
}
\description{
Relative efficiency for cluster randomized trials with binary outcomes
Relative efficiency of varying to equal cluster sizes
}
\examples{
pss.re.crt.bin(m = 60, m.sd = 45, pc = 0.25, pt = 0.15, sigma.u = 0.3)
}
