% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pss.signtest.R
\name{pss.signtest}
\alias{pss.signtest}
\title{Power calculations for sign test (one sample median)}
\usage{
pss.signtest(N = NULL, p = NULL, alpha = 0.05, power = NULL, sides = 2)
}
\arguments{
\item{N}{The sample size.}

\item{p}{The probability of a positive difference when the alternative hypothesis is true.}

\item{alpha}{The significance level or type 1 error rate; defaults to 0.05.}

\item{power}{The specified level of power.}

\item{sides}{Either 1 or 2 (default) to specify a one- or two-sided hypothesis test.}
}
\value{
A list of the arguments (including the computed one).
}
\description{
Power calculations for sign test (one sample median)
}
\examples{
pss.signtest(N = 40, p = 0.7, power = NULL, alpha = 0.05, sides = 1)
}
