% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pss.slr.R
\name{pss.slr}
\alias{pss.slr}
\title{Power calculations for a simple linear regression}
\usage{
pss.slr(
  n = NULL,
  beta10 = 0,
  beta1A = NULL,
  sd.x.sq = NULL,
  sigma.e = NULL,
  alpha = 0.05,
  power = NULL
)
}
\arguments{
\item{n}{The sample size.}

\item{beta10}{The slope regression coefficient under the null hypothesis.}

\item{beta1A}{The slope regression coefficient under the alternative hypothesis.}

\item{sd.x.sq}{The sample variance of the covariate X.}

\item{sigma.e}{The estimated standard deviation of the error terms.}

\item{alpha}{The significance level or type 1 error rate; defaults to 0.05.}

\item{power}{The specified level of power.}
}
\value{
A list of the arguments (including the computed one).
}
\description{
Power calculations for a simple linear regression
}
\examples{
#' # Example 10.3
Yi = beta0 + beta1 * Xi + ei, i = 1,...,n
pss.slr(n = 100, beta10 = 1, beta1A = 1.5, sd.x.sq = 25, sigma.e = 10)
}
