% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pss.t.test.1samp.R
\name{pss.t.test.1samp}
\alias{pss.t.test.1samp}
\title{Power calculations for one sample t tests}
\usage{
pss.t.test.1samp(
  N = NULL,
  delta = NULL,
  sd = 1,
  alpha = 0.05,
  power = NULL,
  sides = 2,
  v = TRUE
)
}
\arguments{
\item{N}{The sample size.}

\item{delta}{muA (the true mean) - mu0 (the mean under the null).}

\item{sd}{The estimated standard deviation; defaults to 1.}

\item{alpha}{The significance level or type 1 error rate; defaults to 0.05.}

\item{power}{The specified level of power.}

\item{sides}{Either 1 or 2 (default) to specify a one- or two- sided hypothesis test.}

\item{v}{Either TRUE for verbose output or FALSE to output computed argument only.}
}
\value{
A list of the arguments (including the computed one).
}
\description{
Power calculations for one sample t tests
}
\examples{
pss.t.test.1samp(N = 36, delta = 4.9 - 5.7, sd = 2, sides = 1)
pss.t.test.1samp(N = 36, delta = 6.3 - 5.7, sd = 2, sides = 1)
pss.t.test.1samp(N = 36, delta = 4.9 - 5.7, sd = 2, sides = 2)
pss.t.test.1samp(delta = 0.6, sd = 1, power = 0.8, sides = 1)
}
