% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pss.t.test.2samp.R
\name{pss.t.test.2samp}
\alias{pss.t.test.2samp}
\title{Power calculations for two sample t tests allowing for unequal sample sizes and/or variances}
\usage{
pss.t.test.2samp(
  n = NULL,
  n.ratio = 1,
  delta = NULL,
  sd = 1,
  sd.ratio = 1,
  df.method = c("welch", "classical"),
  alpha = 0.05,
  power = NULL,
  sides = 2,
  strict = TRUE
)
}
\arguments{
\item{n}{The sample size for group 1.}

\item{n.ratio}{The ratio n2/n1 between the sample sizes of two groups; defaults to 1 (equal group sizes).}

\item{delta}{DeltaA (the true difference mu1 - mu2) - Delta0 (the difference under the null) - delta. See delta.sign for guidance on the sign of delta.}

\item{sd}{The estimated standard deviation for group 1; defaults to 1 (equal standard deviations in the two groups).}

\item{sd.ratio}{The ratio sd2/sd1 between the standard deviations of the two groups.}

\item{df.method}{Method for calculating the degrees of freedom: "welch" (default) or "classical".}

\item{alpha}{The significance level or type 1 error rate; defaults to 0.05.}

\item{power}{The specified level of power.}

\item{sides}{Either 1 or 2 (default) to specify a one- or two- sided hypothesis test.}

\item{strict}{Use strict interpretation in two-sided case; defaults to TRUE.}
}
\value{
A list of the arguments (including the computed one).
}
\description{
Power calculations for two sample t tests allowing for unequal sample sizes and/or variances
}
\examples{
pss.t.test.2samp(n = 50, delta = 2, sd = 5, sides = 1)
pss.t.test.2samp(n = NULL, n.ratio = 2, delta = 0.5, sd = 1, power = 0.8, sides = 2)
pss.t.test.2samp(n = 49, n.ratio = 2, delta = 0.5, sd = 1, power = NULL, sides = 2)
pss.t.test.2samp(n = 25, n.ratio = 3, delta = 3, sd = 4, sd.ratio = 1.5, alpha = 0.025, sides = 1)
pss.t.test.2samp(n = NULL, delta = 0.5, sd = 1, power = 0.8, sides = 2)
}
