% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pss.varym.re.R
\name{pss.varym.re}
\alias{pss.varym.re}
\title{Relative efficiency for cluster randomized or multisite trials
due to varying cluster sizes}
\usage{
pss.varym.re(m, m.sd, icc)
}
\arguments{
\item{m}{The number of subjects per cluster or the mean cluster size (if unequal number of participants per cluster).}

\item{m.sd}{The standard deviation of cluster sizes (provide if unequal number of participants per cluster).}

\item{icc}{The intraclass correlation coefficient. For a multisite trial this is icc1. For a CRT this is the average of the 2 icc's.}
}
\value{
The computed RE.
}
\description{
Relative efficiency for cluster randomized or multisite trials
due to varying cluster sizes
}
\examples{
pss.varym.re(m = 25, m.sd = 15, icc = 0.05)
}
