% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pss.z.test.1samp.R
\name{pss.z.test.1samp}
\alias{pss.z.test.1samp}
\title{Power calculations for one sample z tests}
\usage{
pss.z.test.1samp(
  N = NULL,
  delta = NULL,
  sd = 1,
  alpha = 0.05,
  power = NULL,
  sides = 2,
  strict = TRUE
)
}
\arguments{
\item{N}{The sample size.}

\item{delta}{muA (the true mean) - mu0 (the mean under the null).}

\item{sd}{The estimated standard deviation; defaults to 1.}

\item{alpha}{The significance level or type 1 error rate; defaults to 0.05.}

\item{power}{The specified level of power.}

\item{sides}{Either 1 or 2 (default) to specify a one- or two- sided hypothesis test.}

\item{strict}{Use strict interpretation in two-sided case; defaults to TRUE.}
}
\value{
A list of the arguments (including the computed one).
}
\description{
Power calculations for one sample z tests
}
\examples{
pss.z.test.1samp(N = NULL, delta = 6.5 - 5.7, sd = 2, power = 0.8, sides = 2)
pss.z.test.1samp(N = 40, delta = NULL, sd = 1, power = 0.9, sides = 2)
pss.z.test.1samp(N = NULL, delta = 0.6, sd = 1, power = 0.8, sides = 1)
}
