% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safe.uniroot.R
\name{safe.uniroot}
\alias{safe.uniroot}
\title{Internal-use wrapper for catching errors induced by stats::uniroot}
\usage{
safe.uniroot(f = NULL, interval = NULL)
}
\arguments{
\item{f}{The function for which the root is sought}

\item{interval}{A vector containing the end-points of the interval to be searched for the root}
}
\value{
If no error occurs, returns the solved root. If an error occurs, output a custom message.
}
\description{
Internal-use wrapper for catching errors induced by stats::uniroot
}
\examples{
f <- function(x) x - 3
safe.uniroot(f, c(-10, 10))$root
}
\keyword{internal}
