% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swd_1trt_power.R
\name{swd_1trt_power}
\alias{swd_1trt_power}
\title{Stepped Wedge Design Power Calculation with 1 Treatment}
\usage{
swd_1trt_power(alpha = 0.05, J, m, K, b, d, rho, rho_c, rho_s)
}
\arguments{
\item{alpha}{The significance level or type 1 error rate}

\item{J}{Number of clusters}

\item{m}{The number of individuals in each cluster at each time period}

\item{K}{The number of steps (if one baseline period, then this is periods - 1)}

\item{b}{The number of baseline periods}

\item{d}{The standardized effect size}

\item{rho}{Intraclass correlation coefficient (ICC)}

\item{rho_c}{Cluster autocorrelation (proportion of cluster-level variance that is time invariant)}

\item{rho_s}{Subject autocorrelation (This is 0 in repeated cross sectional designs because people are measured only once)}
}
\value{
power
}
\description{
Calculates power for stepped wedge design with one treatment and one control condition.
}
\examples{
# closed cohort design as in Example 12.5
swd_1trt_power(alpha = 0.05, m = 30, J = 5, K = 5, b = 1, d = 0.4,
               rho = 0.05, rho_c = 0.2, rho_s = 0.4)
# repeated cross sectional design as in Example 12.5
swd_1trt_power(alpha = 0.05, m = 30, J = 5, K = 5, b = 1, d = 0.4,
               rho = 0.05, rho_c = 0.2, rho_s = 0)
}
