% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swd_1trt_ss.R
\name{swd_1trt_ss}
\alias{swd_1trt_ss}
\title{Stepped Wedge Design Sample Size Calculation with 1 Treatment}
\usage{
swd_1trt_ss(alpha, power, m, K, b, d, rho, rho_c, rho_s)
}
\arguments{
\item{alpha}{The significance level or type 1 error rate}

\item{power}{The specified level of power}

\item{m}{The number of individuals in each cluster at each time period}

\item{K}{The number of steps (if one baseline period, then this is periods - 1)}

\item{b}{The number of baseline periods}

\item{d}{The standardized effect size}

\item{rho}{Intraclass correlation coefficient (ICC)}

\item{rho_c}{Cluster autocorrelation (proportion of cluster-level variance that is time invariant)}

\item{rho_s}{Subject autocorrelation (This is 0 in repeated cross sectional designs because people are measured only once)}
}
\value{
Calculated number of clusters required, adjusted number of clusters, and suggested number of clusters that includes the value of K
}
\description{
Calculates sample size requirements for stepped wedge design with one treatment and one control condition.
}
\examples{
# Repeated Cross Sectional Stepped Wedge Sample Size Calculation
# with 3 steps, 1 baseline period, 80\% power, 0.05 significance, 30 ppl per
# cluster, effect size of 0.4, ICC of 0.05, and cluster autocorrelation of 0.2
# Following Example 12.6 in text
swd_1trt_ss(alpha = 0.05, power = 0.8, m = 30, K = 3, b = 1, d = 0.4,
            rho = 0.05, rho_c = 0.2, rho_s = 0)
}
