% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swd_2trt_additive_power.R
\name{swd_2trt_additive_power}
\alias{swd_2trt_additive_power}
\title{Additive Treatments Stepped Wedge Design Power Calculation}
\usage{
swd_2trt_additive_power(
  RhoW,
  ModelChoice,
  n.individuals,
  n.clusters,
  n.periods,
  delta_1,
  delta_2,
  RhoA,
  IAC,
  Sequence_Tx1,
  Sequence_Tx2,
  typeIerror
)
}
\arguments{
\item{RhoW}{The intraclass correlation coefficient, which refers to the correlation between outcomes of two different individuals in the same cluster at the same time.}

\item{ModelChoice}{A string indicating either a repeated cross sectional design ("RCS"), nested exchangeable ("NEC"), Cohort ("Cohort"), or any one of these three designs in their standardized form. Only one of "RCS Standardized", "RCS", "Cohort Standardized", "Cohort", "NEC Standardized", or "NEC" can be specified. The standardized design refers to dividing by the standard deviation of y (the outcome) in the covariance specification.}

\item{n.individuals}{The number of individuals in each condition at one time period.}

\item{n.clusters}{The number of clusters}

\item{n.periods}{The number of time periods.}

\item{delta_1}{The standardized effect size for treatment 1}

\item{delta_2}{The standardized effect size for treatment 2}

\item{RhoA}{The correlation between two observations in the same individual in the same cluster but different time periods. RCS models assume RhoW = RhoA, and Cohort models assume RhoA >= RhoW.}

\item{IAC}{The individual auto-correlation, which is the proportion of the individual-level variance that is time-invariant. IAC = 0 in a RCS.}

\item{Sequence_Tx1}{A vector representing the time points (periods) at which each cluster transitions from control to treatment 1 (ie the first occurence).}

\item{Sequence_Tx2}{A vector representing the time points (periods) at which each cluster transitions from control to treatment 2 (ie the first occurence).}

\item{typeIerror}{The significance level.}
}
\value{
A list with power for detecting treatment effect for treatment 1 and 2
}
\description{
Calculate power for detecting treatment effects in a two-treatment stepped wedge design with additive treatment effects.
We only consider complete designs. This function is based on Power Analysis for Stepped Wedge Trials with Two Treatments published in Stat Med in 2022 by Phillip Sundin and Catherine Crespi. This manuscript uses the normal distribution and the Wald Test to derive power.
}
\examples{
## For example, the 3 cluster, 4 period SWD,
##     0 1 1   1
##     0 0 1+2 1+2
##     0 0 2   1+2
## where 0 denotes control condition, 1 denotes the condition with only
## treatment 1, 2 denote the condition with only treatment 2, and 1+2 denotes the condition
## where a cluster receives both treatment 1 and 2
## would have Sequence1 <- c(2,3,4) and Sequence2 <- c(NA,3,3)
## notice sequence 1 cluster 3 has time period 4 even though time period 3 had treatment 2,
## so it is the first occurrence of trt 1 in this cluster.


# 12-cluster concurrent repeated cross sectional design
# Sequencing has 2 clusters transition to treatment 1 at time 2,
# 2 clusters transition to treatment 1 at time 3, and 2 clusters transition
# to treatment 1 at time 3. These clusters never receive treatment 2
Sequence_Tx1 <- c(2, 2, 3, 3, 4, 4, NA, NA, NA, NA, NA, NA)
# similar sequencing for treatment 2
Sequence_Tx2 <- c(NA, NA, NA, NA, NA, NA, 4, 4, 3, 3, 2, 2)
swd_2trt_additive_power(RhoW = 0.05, ModelChoice = "RCS", RhoA = 0.2, IAC = 0,
               n.individuals = 30, n.clusters = 12, n.periods = 4, delta_1 = 0.4, delta_2 = 0.1,
               Sequence_Tx1 = Sequence_Tx1, Sequence_Tx2 = Sequence_Tx2,
               typeIerror = 0.05)
}
