% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swd_2trt_interaction_power.R
\name{swd_2trt_interaction_power}
\alias{swd_2trt_interaction_power}
\title{Stepped Wedge Design Power Calculation for Interaction Term}
\usage{
swd_2trt_interaction_power(
  RhoW,
  ModelChoice,
  n.individuals,
  n.clusters,
  n.periods,
  delta_1,
  delta_2,
  delta_3,
  RhoA,
  IAC,
  typeIerror,
  Sequence_Tx1,
  Sequence_Tx2
)
}
\arguments{
\item{RhoW}{The intraclass correlation coefficient, which refers to the correlation between outcomes of two different individuals in the same cluster at the same time.}

\item{ModelChoice}{A string indicating either a repeated cross sectional design ("RCS"), nested exchangeable ("NEC"), Cohort ("Cohort"), or any one of these three designs in their standardized form. Only one of "RCS Standardized", "RCS", "Cohort Standardized", "Cohort", "NEC Standardized", or "NEC" can be specified. The standardized design refers to dividing by the standard deviation of y (the outcome) in the covariance specification.}

\item{n.individuals}{The number of individuals in each condition at one time period.}

\item{n.clusters}{The number of clusters}

\item{n.periods}{The number of time periods.}

\item{delta_1}{The standardized effect size for treatment 1}

\item{delta_2}{The standardized effect size for treatment 2}

\item{delta_3}{The standardized effect size for the interaction effect.}

\item{RhoA}{The correlation between two observations in the same individual in the same cluster but different time periods. RCS models assume RhoW = RhoA, and Cohort models assume RhoA >= RhoW.}

\item{IAC}{The individual auto-correlation, which is the proportion of the individual-level variance that is time-invariant. IAC = 0 in a RCS.}

\item{typeIerror}{The significance level.}

\item{Sequence_Tx1}{A vector representing the time points (periods) at which each cluster transitions from control to treatment 1 (ie the first occurence).}

\item{Sequence_Tx2}{A vector representing the time points (periods) at which each cluster transitions from control to treatment 2 (ie the first occurence).}
}
\value{
A matrix with the standard error, statistic, and power for detecting effects for treatment 1, treatment 2, and interaction effects.
}
\description{
Calculate power for detecting interaction effects in a two-treatment stepped wedge design.
We only consider complete designs. This function is based on Power Analysis for Stepped Wedge Trials with Two Treatments published in Stat Med in 2022 by Phillip Sundin and Catherine Crespi. This manuscript uses the normal distribution and the Wald Test to derive power.
}
\examples{
# 8 Cluster repeated cross sectional SWD with main and interaction effects assumed
# to be the same with 0.6 effect size.
swd_2trt_interaction_power(RhoW = 0.2, ModelChoice = "RCS", RhoA = 0.2, IAC = 0,
                n.individuals = 15, n.clusters = 8, n.periods = 5, delta_1 = 0.6,
                delta_2 = 0.6, delta_3 = 0.6, typeIerror = 0.05/3,
                Sequence_Tx1 = c(2,3,2,3,4,5,NA,NA), Sequence_Tx2 = c(NA,NA,2,3,4,5,3,2))
}
