% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swd_2trt_linearcontrast_power.R
\name{swd_2trt_linearcontrast_power}
\alias{swd_2trt_linearcontrast_power}
\title{Stepped Wedge Design Power Calculation for Linear Contrasts of Treatment Effects}
\usage{
swd_2trt_linearcontrast_power(
  RhoW,
  ModelChoice,
  n.individuals,
  n.clusters,
  n.periods,
  delta_12,
  RhoA,
  IAC,
  Sequence_Tx1,
  Sequence_Tx2,
  typeIerror
)
}
\arguments{
\item{RhoW}{The intraclass correlation coefficient, which refers to the correlation between outcomes of two different individuals in the same cluster at the same time.}

\item{ModelChoice}{A string indicating either a repeated cross sectional design ("RCS"), nested exchangeable ("NEC"), Cohort ("Cohort"), or any one of these three designs in their standardized form. Only one of "RCS Standardized", "RCS", "Cohort Standardized", "Cohort", "NEC Standardized", or "NEC" can be specified. The standardized design refers to dividing by the standard deviation of y (the outcome) in the covariance specification.}

\item{n.individuals}{The number of individuals in each condition at one time period.}

\item{n.clusters}{The number of clusters}

\item{n.periods}{The number of time periods.}

\item{delta_12}{The standardized difference between the two treatment effects to be powered on.}

\item{RhoA}{The correlation between two observations in the same individual in the same cluster but different time periods. RCS models assume RhoW = RhoA, and Cohort models assume RhoA >= RhoW.}

\item{IAC}{The individual auto-correlation, which is the proportion of the individual-level variance that is time-invariant. IAC = 0 in a RCS.}

\item{Sequence_Tx1}{A vector representing the time points (periods) at which each cluster transitions from control to treatment 1 (ie the first occurence).}

\item{Sequence_Tx2}{A vector representing the time points (periods) at which each cluster transitions from control to treatment 2 (ie the first occurence).}

\item{typeIerror}{The significance level.}
}
\value{
Power for detecting the difference of treatment effects
}
\description{
Calculate the power for detecting linear contrasts of two treatment effects in a two-treatment stepped wedge design (SWD) with additive treatment effects.
We only consider complete designs. This function is based on Power Analysis for Stepped Wedge Trials with Two Treatments published in Stat Med in 2022 by Phillip Sundin and Catherine Crespi. This manuscript uses the normal distribution and the Wald Test to derive power.
}
\examples{
# Late factorial SWD design with 12 clusters
swd_2trt_linearcontrast_power(RhoW = 0.2, ModelChoice="RCS", IAC = 0,
                      RhoA = 0.2, n.individuals = 15, delta_12 = 0.4,
                      n.clusters = 12, n.periods = 4,
                      Sequence_Tx1 = c(2,2,3,3,4,4,4,4,4,4,4,4),
                      Sequence_Tx2 = c(4,4,4,4,4,4,2,2,3,3,4,4),
                      typeIerror = 0.05)

# Example where there are different potential designs to be considered #
#creating data frame for power as a function of rho_w
PowerTable <- data.frame(RhoW = c(rep(seq(0,0.4,by=0.01),times=3)))
PowerTable$DesignChoice <- c(rep('"Late" Factorial Design, 12 clusters',times=41),
                             rep('"Early" Factorial Design, 10 clusters',times=41),
                             rep("Concurrent Design, 12 clusters",times=41))

PowerTable$n.clusters <- c(rep(12,times=41), rep(10,times=41), rep(12,times=41))

PowerTable$Sequence1 <- c(rep(list(c(2,2,3,3,4,4,4,4,4,4,4,4)),times=41),
                          rep(list(c(2,2,2,3,4,4,3,4,4,4)),times=41),
                          rep(list(c(2,2,3,3,4,4,NA,NA,NA,NA,NA,NA)),times=41))

PowerTable$Sequence2 <- c(rep(list(c(4,4,4,4,4,4,2,2,3,3,4,4)),times=41),
                          rep(list(c(4,4,4,3,4,4,3,2,2,2)),times=41),
                          rep(list(c(NA,NA,NA,NA,NA,NA,4,4,3,3,2,2)),times=41))

PowerTable$Power <- mapply(swd_2trt_linearcontrast_power,
                        RhoW = PowerTable$RhoW,
                        ModelChoice="RCS",
                        IAC = 0, RhoA = PowerTable$RhoW,
                        n.individuals = 15, delta_12 = 0.4,
                        n.clusters = PowerTable$n.clusters,
                        n.periods = 4, typeIerror = 0.05,
                        Sequence_Tx1 = PowerTable$Sequence1,
                        Sequence_Tx2 = PowerTable$Sequence2)

##removing unnecessary columns of clusters
PowerTable <- PowerTable[ ,-which(names(PowerTable) \%in\%
            c("Sequence1","Sequence2","n.clusters"))]
PowerTable
}
