% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t.test.2samp.R
\name{t.test.2samp}
\alias{t.test.2samp}
\title{Power calculations for two sample t tests allowing for unequal sample sizes and/or variances}
\usage{
\method{t}{test.2samp}(
  n1 = NULL,
  n.ratio = 1,
  delta = NULL,
  sd1 = 1,
  sd.ratio = 1,
  df.method = c("welch", "classical"),
  alpha = 0.05,
  power = NULL,
  sides = 2,
  v = TRUE
)
}
\arguments{
\item{n1}{The sample size for group 1.}

\item{n.ratio}{The ratio n2/n1 between the sample sizes of two groups; defaults to 1 (equal group sizes).}

\item{delta}{DeltaA (the true difference mu1 - mu2) - Delta0 (the difference under the null) - margin. See margin.sign for guidance on the sign of the margin.}

\item{sd1}{The estimated standard deviation for group 1; defaults to 1 (equal standard deviations in the two groups).}

\item{sd.ratio}{The ratio sd2/sd1 between the standard deviations of the two groups.}

\item{df.method}{Method for calculating the degrees of freedom: "welch" (default) or "classical".}

\item{alpha}{The significance level or type 1 error rate; defaults to 0.05.}

\item{power}{The specified level of power.}

\item{sides}{Either 1 or 2 (default) to specify a one- or two- sided hypothesis test.}

\item{v}{Either TRUE for verbose output or FALSE to output computed argument only.}
}
\value{
A list of the arguments (including the computed one).
}
\description{
Power calculations for two sample t tests allowing for unequal sample sizes and/or variances
}
\examples{
t.test.2samp(n1 = 50, delta = 2, sd1 = 5, sides = 1)
t.test.2samp(n1 = NULL, n.ratio = 2, delta = 0.5, sd1 = 1, power = 0.8, sides = 2)
t.test.2samp(n1 = 49, n.ratio = 2, delta = 0.5, sd1 = 1, power = NULL, sides = 2)
t.test.2samp(n1 = 25, n.ratio = 3, delta = 3, sd1 = 4, sd.ratio = 1.5, alpha = 0.025, sides = 1)
t.test.2samp(n1 = NULL, delta = 0.5, sd1 = 1, power = 0.8, sides = 2)
}
