% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/es.h.R
\name{es.h}
\alias{es.h}
\title{Cohen h effect size calculation for two proportions}
\usage{
es.h(p1 = NULL, p2 = NULL)
}
\arguments{
\item{p1}{The outcome proportion in group 1.}

\item{p2}{The outcome proportion in group 2.}
}
\value{
A list of the arguments and the h effect size.
}
\description{
Takes as input the outcome proportions in two groups and returns the h effect size
as defined by Cohen (1988).
}
\details{
Cohen J (1988) Statistical Power Analysis for the Behavioral Sciences, 2nd edition.
Lawrence Erlbaum Associates, Hillsdale, New Jersey
}
\examples{
es.h(p1 = 0.8, p2 = 0.6)
}
