% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.format.R
\name{matrix.format}
\alias{matrix.format}
\title{Internal-use function for displaying matrices in function outputs}
\usage{
matrix.format(matrix = NULL, nspaces = 18)
}
\arguments{
\item{matrix}{The matrix to be displayed.}

\item{nspaces}{The number of white spaces before the matrix; defaults to 18.}
}
\value{
The matrix printed as a string with rows separated by newline.
}
\description{
Internal-use function for displaying matrices in function outputs
}
\examples{
matrix <- matrix(c(1, 2, 3, 4), nrow = 2)
matrix.format(matrix = matrix)
}
\keyword{internal}
