% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pss.multisite.bin.R
\name{pss.multisite.bin}
\alias{pss.multisite.bin}
\title{Number of sites for multisite trials with binary outcomes}
\usage{
pss.multisite.bin(
  m = NULL,
  pc = NULL,
  pt = NULL,
  sigma.u = NULL,
  alpha = 0.05,
  power = NULL,
  sides = 2
)
}
\arguments{
\item{m}{The number of subjects per site.}

\item{pc}{The probability of the outcome in the control condition.}

\item{pt}{The probability of the outcome in the treatment condition.}

\item{sigma.u}{Standard deviation of the treatment effect across sites.}

\item{alpha}{The significance level or type 1 error rate; defaults to 0.05.}

\item{sides}{Either 1 or 2 (default) to specify a one- or two- sided hypothesis test.}
}
\value{
A list of the arguments (including the computed power).
}
\description{
Number of sites for multisite trials with binary outcomes
}
\examples{
pss.multisite.bin(m = 30, pc = 0.1, pt = 0.2, sigma.u = 0.4, power = 0.9)
}
