% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pss.multisite.hte.R
\name{pss.multisite.hte}
\alias{pss.multisite.hte}
\title{Power for test of heterogeneity of treatment effect}
\usage{
pss.multisite.hte(
  m = NULL,
  J = NULL,
  ssq.u1 = NULL,
  ssq.e = NULL,
  alpha = 0.05
)
}
\arguments{
\item{m}{The number of subjects per site.}

\item{J}{The number of sites.}

\item{ssq.u1}{The variance of the site-level treatment effects (sigma squared u1) under the alternative.}

\item{ssq.e}{The variance of the observations within sites (sigma squared epsilon) under the alternative.}

\item{alpha}{The significance level or type 1 error rate; defaults to 0.05.}
}
\value{
A list of the arguments (including the computed power).
}
\description{
Power for test of heterogeneity of treatment effect
}
\examples{
pss.multisite.hte(m = 10, J = 30, ssq.u1 = 8, ssq.e = 36)
}
