% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signedrank.R
\name{signedrank}
\alias{signedrank}
\title{Power calculations for signed-rank test}
\usage{
signedrank(N = NULL, ps = NULL, alpha = 0.05, power = NULL, sides = 2)
}
\arguments{
\item{N}{The sample size; number of observations or paired differences.}

\item{ps}{The probability that the sum of two values exceeds zero when the alternative hypothesis is true.}

\item{alpha}{The significance level or type 1 error rate; defaults to 0.05.}

\item{power}{The specified level of power.}

\item{sides}{Either 1 or 2 (default) to specify a one- or two-sided hypothesis test.}
}
\value{
A list of the arguments (including the computed one).
}
\description{
Power calculations for signed-rank test
}
\examples{
signedrank(N = 20, p = 0.87, power = NULL, sides = 2)
}
